Split actions and reducers from store file.
